IssueNo:#I4Y5M6 Description:codeDex告警Fopen无需判断 Sig: SIG_ApplicationFramework Feature or Bugfix: Feature Binary Source: No

+++ services/abilitymgr/src/ability_scheduler_stub.cpp
@@ -544,7 +544,7 @@ int AbilitySchedulerStub::ExecuteBatchInner(MessageParcel &data, MessageParcel &
     }
 
     std::vector<std::shared_ptr<AppExecFwk::DataAbilityResult>> results = ExecuteBatch(operations);
-    int total = results.size();
+    int total = (int)results.size();
     if (!reply.WriteInt32(total)) {
         HILOG_ERROR("AbilitySchedulerStub::ExecuteBatchInner fail to WriteInt32 ret");
         return ERR_INVALID_VALUE;
